Move frontend code around, the presence of NluxAiChat/NluxAiChatProps inside ol-components/src/index.ts was for some reason causing test failures
