+++ frontends/main/package.json
@@ -12,8 +12,11 @@
   "dependencies": {
     "@ebay/nice-modal-react": "^1.2.13",
     "@emotion/cache": "^11.13.1",
+    "@emotion/styled": "^11.11.0",
     "@mitodl/course-search-utils": "3.3.2",
     "@next/bundle-analyzer": "^14.2.15",
+    "@nlux/react": "^2.17.1",
+    "@nlux/themes": "^2.17.1",
     "@remixicon/react": "^4.2.0",
     "@sentry/nextjs": "^8.36.0",
     "@tanstack/react-query": "^4.36.1",
