+++ frontends/main/src/page-components/Nlux-AiChat/mock-send.ts
